move MultiUEMEApp to a more appropriate folder
